Use brooklyn-itest project to signal bundle resolution problems
Add brooklyn-core feature to boot features in apache-brooklyn karaf
assembly. Remove custom setting of org.osgi.framework.system.packages.extra.

Depend on apache-brooklyn assembly in brooklyn-itest. Use
apache-brooklyn as OSGi runtime for AssemblyTest.
